\name{AustralianWine}
\alias{AustralianWine}
\docType{data}
\title{Australian wine sales}
\description{
  Monthly Australian wine sales in thousands of litres
  from Jan 1980 till Jul 1995.
  By wine makers in bottles <= 1 litre.
}
\usage{data(AustralianWine)}
\format{Multivariate time series with 7 columns of length 187 Some columns have missing values at the end.}
\source{
  Sales, Source: Australian Bureau of Statistics 2012, in file: data/wine,
  Description: Monthly Australian wine sales: thousands of litres. By wine makers in bottles <= 1 litre.
}
\keyword{datasets}
