\name{Barbara}
\alias{Barbara}
\docType{data}

\title{Classical `Barbara' image (color, wide)}

\description{
  Classical `Barbara' image (wide version).
  720 x 576 x 3 (color, RGB model), from 0 to 255.
}
\usage{data(Barbara)}

\format{An integer array of dimension 3.}

\source{
  \url{http://www.hlevkin.com/TestImages/classic.htm}
}

\keyword{datasets}
