\name{Mars}
\alias{Mars}
\docType{data}

\title{Webcam image of Mars}

\description{
  Image of Mars obtained by a webcam.
  258 x 275, grayscale, from 0 to 255.
}
\usage{data(Mars)}
\format{A double matrix with integer values.}

\source{
  Thierry, P. Tutorial for IRIS 5.59 (an astronomical images processing software),
  \url{http://www.astrosurf.com/buil/iris/tutorial8/doc23_us.htm}.
  Last updated: 20.12.2005
}
\keyword{datasets}
