\name{MotorVehicle}
\alias{MotorVehicle}
\docType{data}
\title{Total U.S. Domestic and Foreign Car Sales}
\description{
  Monthly series containing total domestic and foreign car sales in the
  USA in thousands, from 1967 till 2010.
}
\usage{data(MotorVehicle)}
\format{A time series of length 541.}
\source{
  U.S. Bureau of Economic Analysis. Table 7.2.5S. Auto and Truck Unit Sales Production Inventories
  Expenditures and {P}rice, 2010.
}
\keyword{datasets}
