\name{USUnemployment}
\alias{USUnemployment}
\docType{data}
\title{U.S. unemployment figures}
\description{
  Monthly U.S. male (16-19 years and from 20 years) and female (16-19 years and from 20 years)
  unemployment figures in thousands from 1948 till 1981.
}
\usage{data(USUnemployment)}
\format{A multivariate time series with 408 observations on 4
  variables. The object is of class 'mts'.}
\source{
  D.F.Andrews and H.M.Herzberg (1985).
}
\keyword{datasets}

