\name{AustralianWine}
\alias{AustralianWine}
\docType{data}
\title{Australian Wine Sales}
\description{
  Monthly Australian wine sales in thousands of litres from Jan 1980 till Jul 1995.
  By wine makers in bottles of less than or equal to 1 litre.
}
\usage{data(AustralianWine)}
\format{A multivariate time series with 187 observations on 7
  variables. The object is of class 'mts'.}
\source{
  Hyndman, R.J. Time Series Data Library, http://data.is/TSDLdemo.
}
\keyword{datasets}
