\name{clplot}
\alias{clplot}

\title{
  Ratio of complete lag vectors given window length
}

\description{
  Function to plot a ratio of complete lag vectors given window length
}

\usage{
clplot(x, \dots)
}

\arguments{
  \item{x}{input series}
  \item{\dots}{further arguments passed to plotting functions}
}

\details{
  Function plots a ratio of complete lag vectors given window length.
}

\seealso{
  \code{\link{Rssa}} for an overview of the package, as well as,
  \code{\link[Rssa:igapfill]{igapfill}},
  \code{\link[Rssa:gapfill]{gapfill}}
  \code{\link[Rssa:summarize.gaps]{summarize.gaps}},
}
